Refactor aggregation in the beam pipeline.
This now performs summation and concatenation of per-chunk statistics in separate stages, rather than doing it all using a single _combining_sum which can be inefficient and cause memory issues.

It also splits up the AggregationState into separate (key, data_array) pairs and aggregates them on a per-key basis, to avoid holding too much data in memory at once. This avoids the need for the more complex CombineMultiStage, however it does require some extra code to help with splitting up and reassembling the AggregationState, and exposes on the Aggregator a public API to aggregate individual leaf DataArrays of the statistics.

PiperOrigin-RevId: 788077599